add tx detail state
