+++ modules/lmscan-frontend/src/main/scala/io/leisuremeta/chain/lmscan/frontend/api/parser/blockDetail.scala
@@ -1,0 +1,20 @@
+package io.leisuremeta.chain.lmscan.frontend
+import io.circe.*, io.circe.generic.semiauto.*
+import io.circe.syntax.*
+import io.circe.parser.*
+
+case class BlockDetail(totalCount: Int, totalPages: Int, payload: List[Tx])
+// case class Tx(
+//     hash: String,
+//     blockNumber: Int,
+//     createdAt: Int,
+//     txType: String,
+//     tokenType: String,
+//     signer: String,
+//     value: Double,
+// )
+
+// object TxParser:
+//   implicit val blocklistDecoder: Decoder[TxList] = deriveDecoder[TxList]
+//   implicit val blockDecoder: Decoder[Tx]         = deriveDecoder[Tx]
+//   def decodeParser(body: String)                 = decode[TxList](body)
