+++ modules/lmscan-frontend/src/main/scala/io/leisuremeta/chain/lmscan/frontend/api/parser/txList.scala
@@ -18,16 +18,3 @@ object TxParser:
   implicit val blocklistDecoder: Decoder[TxList] = deriveDecoder[TxList]
   implicit val blockDecoder: Decoder[Tx]         = deriveDecoder[Tx]
   def decodeParser(body: String)                 = decode[TxList](body)
-
-case class BlockList(totalCount: Int, totalPages: Int, payload: List[Block])
-case class Block(
-    number: Int,
-    hash: String,
-    createdAt: Int,
-    txCount: Int,
-)
-
-object BlockParser:
-  implicit val txlistDecoder: Decoder[BlockList] = deriveDecoder[BlockList]
-  implicit val txDecoder: Decoder[Block]         = deriveDecoder[Block]
-  def decodeParser(body: String)                 = decode[BlockList](body)
